requirements.txt
# ===============================
# Core App & UI
# ===============================
streamlit>=1.30.0

# ===============================
# Deep Learning & CNN
# ===============================
tensorflow>=2.10.0,<2.16.0
keras>=2.10.0

# ===============================
# Computer Vision
# ===============================
opencv-python>=4.7.0.72
Pillow>=9.5.0

# ===============================
# Numerical & Data Handling
# ===============================
numpy>=1.22.0
scipy>=1.9.0
pandas>=1.5.0

# ===============================
# Visualization
# ===============================
matplotlib>=3.6.0
seaborn>=0.12.0

# ===============================
# Face Analysis
# ===============================
deepface>=0.0.79
mtcnn>=0.1.1
retina-face>=0.0.13

# ===============================
# Evaluation Metrics (future-ready)
# ===============================
scikit-learn>=1.2.0

# ===============================
# Optional (performance & stability)
# ===============================
tqdm>=4.64.0
